+++ src/promptflow-tools/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features Added
 - Add "detail" to "Azure OpenAI GPT-4 Turbo with Vision" and "OpenAI GPT-4V" tool inputs.
-- Address the prompt injection issue by roles escape and unescape.
+- Avoid unintended parsing by role process to user inputs in prompt templates.
 
 ## 1.4.0 (2024.03.26)